requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pdfplumber==0.7.5
pdf2image==1.16.3
Pillow==10.4.0
tqdm==4.66.1
regex==2024.9.11
python-multipart==0.0.6
aiofiles==23.1.0
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
